+++ test/configuration_test.dart
@@ -62,6 +62,7 @@ void main() {
         logLevel: 'warning',
         release: 'release-args-config',
         dist: 'dist-args-config',
+        buildPath: 'build_path-args-config',
         webBuildPath: 'web_build_path-args-config',
         commits: 'commits-args-config',
         ignoreMissing: true,
@@ -82,6 +83,7 @@ void main() {
         logLevel: 'debug',
         release: 'release-file-config',
         dist: 'dist-file-config',
+        buildPath: 'build_path-file-config',
         webBuildPath: 'web_build_path-file-config',
         commits: 'commits-file-config',
         ignoreMissing: false,
@@ -108,6 +110,7 @@ void main() {
       expect(sut.logLevel, 'warning');
       expect(sut.release, 'release-args-config');
       expect(sut.dist, 'dist-args-config');
+      expect(sut.buildFilesFolder, 'build_path-args-config');
       expect(
         sut.webBuildFilesFolder,
         fixture.fs.path.join(
@@ -138,6 +141,7 @@ void main() {
         logLevel: 'debug',
         release: 'release-file-config',
         dist: 'dist-file-config',
+        buildPath: 'build_path-file-config',
         webBuildPath: 'web_build_path-file-config',
         commits: 'commits-file-config',
         ignoreMissing: true,
@@ -153,6 +157,7 @@ void main() {
 
       expect(sut.name, 'name-file-config');
       expect(sut.version, 'version-file-config');
+
       expect(sut.uploadDebugSymbols, false);
       expect(sut.uploadSourceMaps, true);
       expect(sut.uploadSources, true);
@@ -164,6 +169,7 @@ void main() {
       expect(sut.logLevel, 'debug');
       expect(sut.release, 'release-file-config');
       expect(sut.dist, 'dist-file-config');
+      expect(sut.buildFilesFolder, 'build_path-file-config');
       expect(
         sut.webBuildFilesFolder,
         fixture.fs.path
@@ -193,6 +199,10 @@ void main() {
       expect(sut.uploadSources, false);
       expect(sut.commits, 'auto');
       expect(sut.ignoreMissing, false);
+      expect(
+        sut.buildFilesFolder,
+        fixture.fs.currentDirectory.path,
+      );
       expect(
         sut.webBuildFilesFolder,
         fixture.fs.path.join(sut.buildFilesFolder, 'build/web'),