+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- Add support for build files folder parameter ([#235](https://github.com/getsentry/sentry-dart-plugin/pull/235))
+
 ### Dependencies
 
 - Bump CLI from v2.31.0 to v2.31.2 ([#234](https://github.com/getsentry/sentry-dart-plugin/pull/234))